requirements.txt
llama-index-llms-ollama==0.5.0
llama-index-readers-file==0.4.1
llama-index-llms-openai==0.3.10
psycopg2-binary==2.9.10
llama-index-embeddings-openai==0.3.1
llama-index-embeddings-huggingface==0.4.0
matplotlib==3.10.0
chainlit==1.3.2
pydantic==2.10.1
streamlit==1.41.1
watchdog==6.0.0